style: Rename modal files to match export names
[Per Faithlife coding guidelines][1].

[1]: https://github.com/Faithlife/CodingGuidelines/blob/3d4f45a7df6443218480d19f77bed5d99ecb9484/sections/javascript/javascript.md#file-names-should-match-export-names
